Fix rooted element traversal.
The upstream Element.nextElement(root) method was failing when root had
no children -- it was returning root's siblings, even though it was supposed
to restrict itself to root's descendants.

This caused Element.getElementsByTagName() and Element.querySelectorAll()
(and other related methods) to return incorrect results when invoked on
elements with no children.

Closes gh #32, gh #31.