Workaround a performance regression in node <= 0.6.
See gh #33.  Node <= 0.6 has a performance issue not present in
node 0.7, 0.8, 0.9, 0.10, or 0.11.  The entire test suite, including
the new test case with a >1M attribute, completes in less than 30s on
all other versions of node, but fails to complete in 10 minutes on
node 0.6.

Reduce the size of the test case if we find that we're running on
node <= 0.6.