Fix slow node insertion when IDs are duplicates
In a real-world case where thousands of nodes with duplicate IDs were
inserted, performance in the node count was O(N^2 log N) due to the
need to sort the byId array after each insertion. So, store duplicates
in an object indexed by nid. When getElementById() is called, do a
linear search for the earliest element in the document order and cache
the result.

This gives O(1) insertion, O(1) deletion and O(N) for an insert or
delete immediately followed by a get, in the pre-existing size of the
list, which is a better time order for all three cases than the existing
code.