Update bundled CSS parser to 0.2.5+domino1.
This is a fork of https://github.com/CSSLint/parser-lib which lives on
the `domino` branch at https://github.com/cscott/parser-lib.  This
update corresponds to the `domino1` tag in that repository.

To update, `ant release` in the checked-out `parser-lib` and then
`cp release/npm/lib/node-parserlib.js $DOMINO/lib/cssparser.js`.