Add additional optional parameter to `domino.createDocument`.
We tested the falsiness of the first parameter to determine whether to
parse the parameter, or just create an empty HTML document.  The result
of parsing the empty string is actually slightly different that the
result of `DOMImplementation.createHTMLDocument("")`.  If the difference
matters to you, you can use the new optional second parameter to
`domino.createDocument` to force the empty string to be parsed as
a document, even though it is falsy.