Add basic support for SVG api.
This PR adds SVG elements, as discussed briefly in #81. The key feature
implemented here is that
`document.createElementNS('http://www.w3.org/2000/svg', 'svg')`
now returns a real `SVGSVGElement`, not a plain `Element`, and the
`SVGSVGElement` defines a `createSVGRect` method.  This just happens to
be how Leaflet does SVG detection:
https://github.com/Leaflet/Leaflet/blob/master/src/layer/vector/SVG.js#L225.

This commit includes a very basic test to check that core
functionality is working, and some refactoring to share the element
definition code between HTML and SVG elements. That element definition
is mostly as it was, but now parameterized so it can take different
fallback super constructors (`HTMLElement` vs `SVGElement`), and adds the
resulting elements to the separate lists of HTML/SVG elements.

Closes: #81, #82