Fix W3C test harness to work on node >= 0.11.7.
We were running all the tests in the same context previously.  The tests
all define a `runTest` function --- but the redefinition of that function
in the same context doesn't have any effect after node 0.11.7, with the
result that these versions of node effectively ran the first test case
over and over again, instead of running all the individual test cases.