Make key properties writable to allow polyfilling
Many polyfills rewrite built-ins on the DOM to add new functionality and
support new upcoming specs. In the browser, this is fine - DOM
properties are reliably writable. In Domino they weren't.

The specific changes here are to support the custom-elements polyfill
  https://github.com/webcomponents/custom-elements/blob/master/src/custom-elements.js
but cover the most popular and widely applicable methods. This should
probably be extended in future to cover the whole API.

Writability is configurable.  If you wish to lock down the DOM API,
set the global `__domino_frozen__` to `true` before your first
`require('domino')`.

Fixes: #89