Harmonize expectations of hasAttribute and removeAttribute.
hasAttribute checks for properties via `in this._attrsByQName` which
will check if the property exists in the object at all. But
removeAttribute clears the property via
`this._attrsByQName[qname] = undefined` which doesn't clear the property
as far as the `in` operator is concerned.

So, don't use `in` to test for the presence of attributes.

Fixes: #90, #91