Fix unescape in attribute selectors; remove nonstandard "i" mode.
Zest seems to implement a non-standard "ignore case" version of
attribute selectors, like:
  a[href=foo i]
which was supposed to do a case-insensitive match against the value
`foo`.  This isn't in any CSS standard that I'm aware of, so disable
it.

Also fix the unescape mechanism in values, accounting for different
behavior of quoted values and IDENT values.  The CSS syntax also
allows NUMBER, PERCENTAGE, DIMENSION, and some other stuff here; don't
try to unescape that.

Fixes: #95