Implement ChildNode#after, ChildNode#before, and ChildNode#replaceWith.
In the process, tweak Node#insertBefore, Node#replaceChild, and
Node#removeChild for proper hierarchy checks.

The latest version of the DOM standard adopts nodes rather than failing
with a hierarchy check.  Some other constraints are subtly different as
well.  Reimplement a strict check based on the spec text, fixing quite
a few of our web-platform-tests test cases in the process.