Update `mocha` dependency to 4.0.x
Updating to the latest version of mocha means we are no longer testing
on pre-node-4 releases.  See:
  https://boneskull.com/mocha-v4-nears-release/

And in a more-opinionated manner:
  https://medium.com/@mikeal/stop-supporting-old-releases-70cfa0e04b0c

We do not yet have any code in domino that would break when run on
older versions of node, but I expect we will gradually modernize the
code base after the next release.