Release domino 2.0.0.
We are bumping the major version number because in this release series we
are dropping support for node versions prior to node 4.  This initial
release does not contain any backward-incompatible code other than in
the testing framework, but the plan is to gradually modernize the code
base during the 2.0.x series using features present in node 4.