ContainerNode removeChildren(): ensure children are unlinked from parent
See the issue ticket #129 for a description of what happens here, but in
essence, the problem is that removeChildren() tries to null each kid
parentNode then navigate to the following sibling with nextSibling, but
nextSibling will never return anything if it doesn't have a parent.

So we track nextSibling first, then null the parentNode.