Tweak how `Document#createElement()` handles non-HTML documents
Attempt to match spec behavior more closely for `createElement()` and
`createElementNS()` on non-HTML documents.  We don't really support
XML and XHTML parsing modes, but we've got enough support to run some
of the web-platform-tests suite.

Also cache the uppercased tag names, since converting strings to
uppercase can be expensive when you're doing it for every tag in
the document.