Ensure NodeIterator#filter/TreeWalker#filter reflects original value
The original implementation wrapped the `filter` argument to the
constructor in a `NodeFilter` object.  This isn't how the DOM IDL
says this should work: the filter property should reflect the original
JS object passed (whether a function or an object), and the "call a
user object's operation" mechanism is used to invoke the notional
`NodeFilter#acceptNode`:
  https://heycam.github.io/webidl/#call-a-user-objects-operation

Reflecting the original constructor parameter makes most of the
newly-added dom/traversal tests from web-platform-tests pass.