TreeWalker spec-compliance: filter & exception types, read-only properties
Ensure that most TreeWalker properties are read-only, as specified in
the IDL.  Used the `utils` helpers to ensure that the exceptions thrown
were of the correct types as well.  Finally, follow the "to filter a node"
steps more closely and cast the result to a number to ensure that we handle
filter functions returning booleans correctly.