Preserve element prefix in localName during "insert a foreign element"
The "insert a foreign element" operation in the HTML parsing algorithm
differs from the standard `Document#createElementNS()` operation: it
explicitly sets the prefix to `null`, and creates a `localName` which
may contain `:` characters.  Duplicate this behavior so that prefixes
aren't lost when an SVG element is inserted, since the HTML serialization
algorithm otherwise suppresses prefixes on SVG elements.

This is needed for DOMPurify (#102).