Avoid megamorphic call-site for `Node#_serializeOne()`
This adds a `NodeUtils` helper object, and moves the
`Node#_serializeOne()` there, to avoid the lookup of the
`_serializeOne()` method on the prototype in various places. The problem
with code like

```js
str = this._serializeOne(parent);
```

is that it's called with various different shapes for `this` (i.e.
different `Node` subclass instances), and thus this property access is
going to be *MEGAMORPHIC* in any realistic application. In Node.js this
will hit the so-called *megamorphic stub cache* in V8, which is a global
resource used to cache megamorphic lookups. Having lot's of contention
on this global resource is going to affect various different parts of
the application (i.e. it's a non-local effect on overall performance).
So anything that can be done to reduce the number of megamorphic
property accesses is going to eventually help with performance. The
problem can be avoided easily for `_serializeOne()` by moving the method
off of the `Node.prototype` and into a separate `NodeUtils` helper
object instead.

On a simple Angular Universal [prerender test case][1] (using domino)
this can greatly reduce the number of megamorphic stub cache misses -
i.e. the times when V8 hits the super slow path and has to redo the
full property lookup via the prototype chain - for `LoadIC`s from around
1518983 to 1255967, which corresponds to a **17%** reduction.

Similar improvements could be made for `Node#_countChildrenOfType()`,
`Node#_ensureInsertValid()` and `Node#_insertOrReplace()`, although
that might be less impactful overall.

[1]:
https://gist.githubusercontent.com/bmeurer/5b9480ef1a74c5187180193abc73dcd4/raw/fc5d9a01a6b7fe841efa1c7773ae55e70435d16a/prerender.js